requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
streamlit
extra-streamlit-components==0.1.71
requests

langchain==0.0.208
langchain-community==0.0.38

faiss-cpu==1.7.4

sentence-transformers==2.2.2
transformers==4.30.2
torch==2.0.1

numpy==1.24.4

pydantic==1.10.13
python-multipart

bcrypt==4.1.1
python-jose[cryptography]==3.3.0
SQLAlchemy==2.0.23
python-dotenv==1.0.0